C.D.
I am pushing this to my branch since it isn't completely functional, i
am still working on it detecting was and being able to move around in
the map. I would like to know where we get the players starting
coordinates because that is affecting the tests I am trying to run.
Thanks.