mr bowe to the rescue
after days of frustration i asked for help. this is about the only way
to get it to work. where rungame is essentially our main we have to have
the player movement accessible and not independently running somewhere
else so that we can have collision dectection and toucheswall able to
communicate to it