star movement + collision
Added star movement with player interaction. Also added star collision
against a wall and another star. Had to go about it without modifying
star class because issues arose that would cause a big change in the
Star vector in level.h and level.cpp. Only thing left is restricting
player movement from going on top of a star when a star is next to a
wall or when next to another star.